Add Text Exercise Assessment View to automatic Screenshot
